requirements.txt
Pillow
colorlog
matplotlib
numpy==1.16.6
tensorflow>=1.0.0,<2.0.0